Add more namespaces, and refactor client
