+++ lib/amadeus/client.rb
@@ -1,11 +1,13 @@
 require 'logger'
 require 'amadeus/client/validator'
+require 'amadeus/client/namespaces'
 
 module Amadeus
   # The Amadeus client library for accessing
   # the travel APIs.
   class Client
     include Validator
+    include Namespaces
 
     # The API key used to authenticate against the API
     attr_reader :api_key
@@ -31,36 +33,6 @@ module Amadeus
       @api_secret = initialize_required(:api_secret, options)
       @logger = initialize_optional(:logger, options, Logger.new(STDOUT))
       @logger.level = initialize_optional(:log_level, options, Logger::WARN)
-    end
-
-    # The namespace for the checkin links and locations APIs:
-    #
-    #   amadeus.reference_data.urls.checkin_links
-    #   amadeus.reference_data.locations
-    #
-    def reference_data
-      Amadeus::Client::ReferenceData.new(self)
-    end
-
-    # The namespace for the shopping APIs:
-    #
-    #   amadeus.shopping.flight_destinations
-    #   amadeus.shopping.flight_offers
-    #   amadeus.shopping.flight_dates
-    #   amadeus.shopping.hotel_offers
-    #   amadeus.shopping.hotels
-    #
-    def shopping
-      Amadeus::Client::Shopping.new(self)
-    end
-
-    # The namespace for the travel analytics APIs:
-    #
-    #   amadeus.travel.analytics.air_traffics
-    #   amadeus.travel.analytics.fare_searches
-    #
-    def travel
-      Amadeus::Client::Travel.new(self)
     end
   end
 end
